fix(core): align http host registry and contracts
